[LIB-787] reorganize tests a little ;)
